requirements.txt
mako==1.2.4
pyjwt==2.6.0
albumentations==1.3.0
#cucim==23.4.1
einops==0.6.1
future==0.18.2
geojson==3.0.0
jupyterlab==4.0.2
ipython-genutils==0.2.0
keras==2.12.0
matplotlib==3.7.1
numpy>1.22,<1.24
opencv_python_headless==4.5.4.58
openslide_python==1.2.0
pandarallel==1.6.5
pandas==1.5.3
Pillow==9.5.0
pydantic==1.10.4
PyYAML==6.0
rasterio==1.3.5.post1
schema==0.7.5
scikit-image==0.19.3
scikit-learn==1.2.1
scikit-survival==0.21.0
scipy==1.10.0
seaborn==0.12.2
SimpleITK==2.2.1
Shapely==1.8.5.post1
sqlalchemy==2.0.17
tabulate==0.9.0
tensorflow==2.12.0
tqdm==4.65.0
ujson==5.8.0
wandb==0.15.4
torchinfo==1.8.0
torchmetrics==0.11.4
